+++ statsd/aggregator.go
@@ -31,11 +31,11 @@ type metricAggregatorStats struct {
 // Incoming metrics should be sent to the MetricQueue channel.
 type MetricAggregator struct {
 	sync.Mutex
-	ExpiryInterval    time.Duration     // How often to expire metrics
-	FlushInterval     time.Duration     // How often to flush metrics to the sender
-	LastFlush         time.Time         // Last time the metrics where aggregated
-	MaxWorkers        int               // Number of workers to metrics queue
-	MetricQueue       chan types.Metric // Queue on which metrics are received
+	ExpiryInterval    time.Duration      // How often to expire metrics
+	FlushInterval     time.Duration      // How often to flush metrics to the sender
+	LastFlush         time.Time          // Last time the metrics where aggregated
+	MaxWorkers        int                // Number of workers to metrics queue
+	MetricQueue       chan *types.Metric // Queue on which metrics are received
 	PercentThresholds []float64
 	Senders           []backend.MetricSender // The sender to which metrics are flushed
 	Stats             metricAggregatorStats
@@ -50,7 +50,7 @@ func NewMetricAggregator(senders []backend.MetricSender, percentThresholds []flo
 	a.LastFlush = time.Now()
 	a.ExpiryInterval = expiryInterval
 	a.Senders = senders
-	a.MetricQueue = make(chan types.Metric, maxQueueSize*10) // we are going to receive more metrics than messages
+	a.MetricQueue = make(chan *types.Metric, maxQueueSize*10) // we are going to receive more metrics than messages
 	a.MaxWorkers = maxWorkers
 	a.PercentThresholds = percentThresholds
 	a.Counters = types.Counters{}
@@ -324,7 +324,7 @@ func (a *MetricAggregator) receiveSet(name, tags string, value string, now time.
 }
 
 // receiveMetric is called for each incoming metric on MetricChan
-func (a *MetricAggregator) receiveMetric(m types.Metric, now time.Time) {
+func (a *MetricAggregator) receiveMetric(m *types.Metric, now time.Time) {
 	a.Lock()
 	defer a.Unlock()